+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ### 4.4-SNAPSHOT
   Bugs
   * Fix #1690: Endpoints is always pluralized
+  * Fix #1684: Fixed URL resolution algorithm for OpenShift resources without API Group name
 
   Improvements
   * Fix #1650: Introduced `kubernetes.disable.autoConfig` system property to disable auto configuration in Config
+  * Fix #1661: Remove generic parameter from KubernetesResource
 
   Dependency Upgrade