+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   Improvements
   * Fix #1650: Introduced `kubernetes.disable.autoConfig` system property to disable auto configuration in Config
   * Fix #1661: Remove generic parameter from KubernetesResource
+  * Improved OpenShiftOperation.wrap method performance
   * RawCustomResourceOperationsImpl#makeCall now closes the created Response object
 
   Dependency Upgrade